Force Firebird to use UpperCaseNamingStategy
FireBird works better if we use table and field names in upper case.
If we have a poco named User, we can easily get in trouble with a query
like this:
```
select * from User;
```

Forcing the FirebirdOrmLiteDialectProvider to use the
UpperCaseNamingStrategy, this problem doesn't exist.

```
public FirebirdOrmLiteDialectProvider(bool compactGuid)
{
...
NamingStrategy = new UpperCaseNamingStrategy();
}
```